requirements.txt
adal==1.2.3
applicationinsights==0.11.9
atomicwrites==1.4.0
attrs==19.3.0
azure-common==1.1.25
azure-core==1.5.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-network==10.2.0
azure-mgmt-resource==9.0.0
azure-mgmt-storage==10.0.0
azure-storage-blob==12.3.1
azureml-automl-core==1.7.0.post1
azureml-automl-runtime==1.7.0.post2
azureml-core==1.7.0
azureml-dataprep==1.7.0
azureml-dataprep-native==14.2.1
azureml-defaults==1.7.0
azureml-explain-model==1.7.0
azureml-interpret==1.7.0.post1
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline==1.7.0
azureml-pipeline-core==1.7.0.post1
azureml-pipeline-steps==1.7.0
azureml-sdk==1.7.0.post1
azureml-telemetry==1.7.0
azureml-train==1.7.0
azureml-train-automl==1.7.0
azureml-train-automl-client==1.7.0.post1
azureml-train-automl-runtime==1.7.0.post1
azureml-train-core==1.7.0
azureml-train-restclients-hyperdrive==1.7.0
backports.tempfile==1.0
backports.weakref==1.0.post1
boto==2.49.0
boto3==1.13.17
botocore==1.16.17
Brotli==1.0.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
configparser==3.7.4
contextlib2==0.6.0.post1
coverage==5.1
cryptography==2.9.2
Cython==0.29.19
dash==1.13.3
dash-bootstrap-components==0.10.2
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
dill==0.3.1.1
distro==1.5.0
docker==4.2.0
docutils==0.15.2
dotnetcore2==2.1.14
entrypoints==0.3
fire==0.3.1
flake8==3.7.9
Flask==1.0.3
Flask-Compress==1.5.0
fusepy==3.0.1
future==0.18.2
gensim==3.8.3
gunicorn==19.9.0
idna==2.9
importlib-metadata==1.6.0
inference-schema==1.0.2
interpret-community==0.11.2
interpret-core==0.1.21
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json-logging-py==0.2
JsonForm==0.0.2
jsonpickle==1.4.1
jsonschema==3.2.0
JsonSir==0.0.2
keras2onnx==1.6.1
liac-arff==2.4.0
lightgbm==2.3.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
msal==1.3.0
msal-extensions==0.1.3
msrest==0.6.14
msrestazure==0.6.3
ndg-httpsclient==0.5.1
nimbusml==1.7.1
numpy==1.16.2
oauthlib==3.1.0
onnx==1.6.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.0.0
packaging==20.4
pandas==0.23.4
pathspec==0.8.0
patsy==0.5.1
plotly==4.8.1
pluggy==0.13.1
pmdarima==1.1.1
portalocker==1.7.0
protobuf==3.12.1
psutil==5.7.0
py==1.8.1
py-cpuinfo==5.0.0
pyarrow==0.17.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
Python-EasyConfig==0.1.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
Resource==0.2.1
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-learn==0.20.3
scipy==1.4.1
SecretStorage==3.1.2
shap==0.34.0
six==1.15.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
smart-open==1.9.0
statsmodels==0.10.2
termcolor==1.1.0
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.3
websocket-client==0.57.0
Werkzeug==0.16.1
wrapt==1.11.1
zipp==3.1.0
